HabitatUL/src/test/java/ca/ulaval/glo4003/underwriting/domain/quote/form/identity/UniversityProfileTest.java
package ca.ulaval.glo4003.underwriting.domain.quote.form.identity;

import ca.ulaval.glo4003.helper.quote.form.UniversityProfileGenerator;
import org.junit.Test;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

public class UniversityProfileTest {
  private UniversityProfile subject;

  @Test
  public void checkingIfFormIsFilled_withUnfilledForm_shouldReturnFalse() {
    subject = new UniversityProfile(null, null, null);

    assertFalse(subject.isFilled());
  }

  @Test
  public void checkingIfFormIsFilled_withFilledForm_shouldReturnTrue() {
    subject = UniversityProfileGenerator.createUniversityProfile();

    assertTrue(subject.isFilled());
  }
}
